+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Edge
 * Replace --quiet with -s in as it's no longer supported in newer GIT versions
 * Add support for ENV passing to the partials
+* Add support for single file processing and env passing in the extractor (release)
 
 ## Version 0.6.5
 * Allow disabling of URL relativizing in the extractor with `release.extract :url_relativize => false`